requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
numpy==1.26.0
pillow==10.1.0
pandas==2.1.0
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.13.0
